🏗️ Architect: Modularized Observation class
Decomposed the monolithic `apts/observations/base.py` "God Object" into a focused MixIn architecture.
- Extracted celestial catalog logic to `catalogs.py`.
- Moved HTML export and notification logic to `html.py`.
- Relocated astronomical event calculation to `events.py`.
- Centralized plotting dispatcher logic in `plotting.py`.
- Refactored `Observation.__init__` and `to_html` to reduce cyclomatic complexity.

This structural improvement adheres to the Single Responsibility Principle and enhances maintainability.
